+++ README.md
@@ -14,10 +14,12 @@ if you were running commands directly in the terminal.
 
 I personally use this for running [gulp](https://github.com/gulpjs/gulp) tasks,
 since I got used to using npm scripts and their ability to directly run terminal
-commands very easily. Since it returns a `Promise<SpawnSyncReturns<Buffer>>`,
-[it can be directly used in gulp](https://gulpjs.com/docs/en/getting-started/async-completion#returning-a-promise),
-see the [`gulpfile.babel.ts`](https://github.com/dbpiper/terminal-spawn/blob/master/gulpfile.babel.ts)
-in this project for an example.
+commands very easily. Since it returns _both_ a [`Promise`][promise] and a
+[`ChildProcess`][child_process] it can [easily be used with gulp.][gulp-async-completion]
+
+The [promise][promise] resolves with the same information as [`spawnSync`][spawn-sync-returns]
+and does so once the [`close` event has been received][event-close] and thus you
+can await the promise to resolve if you wish to ensure it completed.
 
 This project uses [TypeScript][typescript] and thus has types for it exported,
 so it works well in that environment. However, it also works just fine with
@@ -46,7 +48,7 @@ import terminalSpawn from 'terminal-spawn';
 
 // execute inside of IIAFE since we can't use top-level await
 (async () => {
-  const subprocess = await terminalSpawn('echo "hello world!"');
+  const subprocess = await terminalSpawn('echo "hello world!"').promise;
 
   if (subprocess.status === 0) {
     console.log('everything went well!');
@@ -56,36 +58,85 @@ import terminalSpawn from 'terminal-spawn';
 })();
 ```
 
+### To spawn a task which never terminates, killing it and ensuring it was killed
+
+```typescript
+import terminalSpawn from 'terminal-spawn';
+
+// execute inside of IIAFE since we can't use top-level await
+(async () => {
+  const subprocessSpawn = terminalSpawn(`
+    while true
+    do
+      echo "hello world!"
+      sleep 0.25
+    done
+  `);
+  // wait for 500 ms to pass...
+  const timeToWait = 500;
+  await new Promise((resolve, _reject) =>
+    setTimeout(() => {
+      resolve();
+    }, timeToWait),
+  );
+  subprocessSpawn.process.kill();
+  // subprocess.signal should be 'SIGTERM'
+  const subprocess = await subprocessSpawn.promise;
+})();
+```
+
 ## API
 
 ### terminalSpawn(command, options)
 
-return type: `Promise<SpawnSyncReturns<Buffer>>`
+return type:
+
+```ts
+  {
+    promise: Promise<SpawnSyncReturns<Buffer>>
+    process: ChildProcess
+  }
+```
 
 Executes the command inside of Node.js as if it were run in the shell. If
 command is an array then the commands will be run in series/sequentially.
 
-The result is a [`Promise`][promise] which has the same structure/type as the
-[return value of the synchronous version of `child_process.spawn`](spawn-sync-returns).
+The result is an object which contains _both_ a [`Promise`][promise] which has
+the same structure/type as the [return value of the synchronous version of `child_process.spawn`][spawn-sync-returns].
+and also a ['ChildProcess`][child_process]. Each of these are useful in certain
+circumstances, for example you need the process reference if you want to kill
+an infinite process. You may want to use the promise to check status codes
+or termination signals to verify that the process actually ended and how.
 
 ### terminalSpawnParallel(command, options)
 
-return type: `Promise<SpawnSyncReturns<Buffer>>`
+return type:
+
+```ts
+  {
+    promise: Promise<SpawnSyncReturns<Buffer>>
+    process: ChildProcess
+  }
+```
 
 Executes the command inside of Node.js as if it were run in the shell, if
 command is an array then the commands will be run in parallel rather than
 in series/sequentially.
 
-The result is a [`Promise`][promise] which has the same structure/type as the
-[return value of the synchronous version of `child_process.spawn`](spawn-sync-returns).
+The result is an object which contains _both_ a [`Promise`][promise] which has
+the same structure/type as the [return value of the synchronous version of `child_process.spawn`][spawn-sync-returns].
+and also a ['ChildProcess`][child_process]. Each of these are useful in certain
+circumstances, for example you need the process reference if you want to kill
+an infinite process. You may want to use the promise to check status codes
+or termination signals to verify that the process actually ended and how.
 
 ### command
 
 type: `string` or `string[]`
 
-The command will be run using the shell and the output will be redirected to the shell.
-This means that it will essentially function as if you ran it directly in a
-shell such as `/bin/sh`, but inside of Node.js.
+The command will be run using the shell and the output will be redirected to the
+shell. This means that it will essentially function as if you ran it directly in
+a shell such as `/bin/sh`, but inside of Node.js.
 
 If command is an array then all of the commands in the array will be executed:
 either in series or in parallel, depending on the function. The default is to
@@ -108,16 +159,16 @@ By default they are:
   }
 ```
 
+Which allows `terminalSpawn` to act like a terminal. However, if you wanted the
+nice argument passing of terminalSpawn, e.g. `'echo "hello world!"` without
+**actually** using the terminal, then you could disable this using `options`.
+
 The API for options is designed to be as user-friendly as possible thus,
 it assumes that you want to keep the terminal-like behavior, but may want
 to change other options such as using `cwd`. To support this the user-provided
 options are added to the default options, rather than always overwriting them
 (aka. set union). However, if you explicitly specify a a default command such
-as `stdio` then it *will* be overwritten.
-
-Which allows `terminalSpawn` to act like a terminal. However, if you wanted the
-nice argument passing of terminalSpawn, e.g. `'echo "hello world!"` without
-**actually** using the terminal, then you could disable this using `options`.
+as `stdio` then it _will_ be overwritten.
 
 However, it should be noted that if you pass the option `shell: false` then
 many features such as multiple commands run in series or parallel will not work
@@ -129,5 +180,8 @@ due to reliance on running in a shell.
 
 [promise]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise
 [spawn-sync-returns]: https://nodejs.org/api/child_process.html#child_process_child_process_spawnsync_command_args_options
+[child_process]: https://nodejs.org/api/child_process.html#child_process_class_childprocess
 [child_process.spawn]: https://nodejs.org/api/child_process.html#child_process_child_process_spawn_command_args_options
 [typescript]: https://www.typescriptlang.org/
+[gulp-async-completion]: https://gulpjs.com/docs/en/getting-started/async-completion
+[event-close]: https://nodejs.org/api/child_process.html#child_process_event_close
