+++ src/index.ts
@@ -1,124 +1,41 @@
+import { ChildProcess, SpawnOptions, SpawnSyncReturns } from 'child_process';
 import {
-  ChildProcess,
-  spawn,
-  SpawnOptions,
-  SpawnSyncReturns,
-} from 'child_process';
+  terminalSpawnParallelProcess,
+  terminalSpawnProcess,
+  terminalSpawnPromiseWrapper,
+} from './util';
 
-export type SpawnPromiseReturns = SpawnSyncReturns<Buffer>;
+export type SpawnPromiseReturns = Promise<SpawnSyncReturns<Buffer>>;
 export type Command = string | string[];
-export type TerminalSpawn = (
+
+export interface TerminalSpawnReturns {
+  process: ChildProcess;
+  promise: SpawnPromiseReturns;
+}
+
+const terminalSpawn = (
   command: Command,
   options?: SpawnOptions,
-) => ChildProcess;
-
-/* *****************************************************************************
- * Private Variables
- **************************************************************************** */
-
-const _spawnOptions: SpawnOptions = {
-  stdio: 'inherit',
-  shell: true,
-};
-
-const _spawnWithStringParser = (shellCommand: string) => {
-  const shellCommandArray = shellCommand.split(' ');
+): TerminalSpawnReturns => {
+  const subprocess = terminalSpawnProcess(command, options);
+  const promise = terminalSpawnPromiseWrapper(subprocess);
   return {
-    command: shellCommandArray[0],
-    args: shellCommandArray.slice(1),
+    promise,
+    process: subprocess,
   };
 };
 
-const _terminalSpawnPromiseWrapper = (
-  command: Command,
-  spawnFunction: TerminalSpawn,
-  options?: SpawnOptions,
-) =>
-  new Promise<SpawnPromiseReturns>((resolve, _reject) => {
-    const subprocess = spawnFunction(command, options);
-    let stdin = '';
-    let stdout = '';
-    let stderr = '';
-    let error = {
-      name: '',
-      message: '',
-    };
-
-    if (subprocess.stdin) {
-      subprocess.stdin.on('data', (chunk: Buffer) => {
-        stdin += chunk.toString();
-      });
-    }
-
-    if (subprocess.stdout) {
-      subprocess.stdout.on('data', (chunk: Buffer) => {
-        stdout += chunk.toString();
-      });
-    }
-
-    if (subprocess.stderr) {
-      subprocess.stderr.on('data', (chunk: Buffer) => {
-        stderr += chunk.toString();
-      });
-    }
-
-    subprocess.on('error', (err: Error) => {
-      error = err;
-    });
-
-    subprocess.on('close', (code: number, signal: string) => {
-      resolve({
-        error,
-        signal,
-        pid: process.pid,
-        output: [stdin, stdout, stderr],
-        stdout: Buffer.from(stdout),
-        stderr: Buffer.from(stderr),
-        status: code,
-      });
-    });
-  });
-
-const _terminalSpawnProcess: TerminalSpawn = (
+const terminalSpawnParallel = (
   command: Command,
   options?: SpawnOptions,
-) => {
-  let terminalCommand = '';
-  if (Array.isArray(command)) {
-    terminalCommand = command.join(' && ');
-  } else {
-    terminalCommand = command as string;
-  }
-  const commandObj = _spawnWithStringParser(terminalCommand);
-  const spawnOptions: SpawnOptions = { ..._spawnOptions, ...options };
-  return spawn(commandObj.command, commandObj.args, spawnOptions);
+): TerminalSpawnReturns => {
+  const subprocess = terminalSpawnParallelProcess(command, options);
+  const promise = terminalSpawnPromiseWrapper(subprocess);
+  return {
+    promise,
+    process: subprocess,
+  };
 };
-
-const _terminalSpawnParallelProcess: TerminalSpawn = (
-  command: Command,
-  options?: SpawnOptions,
-) => {
-  let terminalCommand = '';
-  if (Array.isArray(command)) {
-    terminalCommand = command.join(' & ');
-    terminalCommand += '; wait';
-  } else {
-    terminalCommand = command as string;
-  }
-  const commandObj = _spawnWithStringParser(terminalCommand);
-  const spawnOptions: SpawnOptions = { ..._spawnOptions, ...options };
-  return spawn(commandObj.command, commandObj.args, spawnOptions);
-};
-
-/* *****************************************************************************
- * Public Variables
- **************************************************************************** */
-
-const terminalSpawn = (command: Command, options?: SpawnOptions) =>
-  _terminalSpawnPromiseWrapper(command, _terminalSpawnProcess, options);
-
-const terminalSpawnParallel = (command: Command, options?: SpawnOptions) =>
-  _terminalSpawnPromiseWrapper(command, _terminalSpawnParallelProcess, options);
 
 export default terminalSpawn;
